LAB 15 - refreshing content
