+++ Allfiles/Labs/15-Remotely monitor and control devices with Azure IoT Hub/Final/CheeseCaveOperator/CheeseCaveOperator.csproj
@@ -1,0 +1,13 @@
+<Project Sdk="Microsoft.NET.Sdk">
+
+  <PropertyGroup>
+    <OutputType>Exe</OutputType>
+    <TargetFramework>netcoreapp3.1</TargetFramework>
+  </PropertyGroup>
+
+  <ItemGroup>
+    <PackageReference Include="Microsoft.Azure.Devices" Version="1.*" />
+    <PackageReference Include="Microsoft.Azure.EventHubs" Version="4.*" />
+  </ItemGroup>
+
+</Project>
